Move member_fields.js and form_fields/member.js into shared.
