+++ src/admin/dumb_components/view_member.js
@@ -1,7 +1,7 @@
 const React = require('react')
 const { reduxForm } = require('redux-form')
-const { validate, required } = require('../form_fields/member.js')
-const MemberFields = require('./member_fields.js')
+const { validate, required, read_only } = require('../../shared/form_fields/member.js')
+const MemberFields = require('../../shared/dumb_components/member_fields.js')
 
 module.exports = (
   props
@@ -10,6 +10,7 @@ module.exports = (
     {...props}
     required={required}
     buttons_first={true}
+    read_only={read_only}
   />
 
 const ViewMember = reduxForm(
