+++ src/user/components/my_details/account_details.js
@@ -1,14 +1,14 @@
 import React from 'react'
-import MemberFields from '../../../admin/dumb_components/member_fields.js'
+import MemberFields from '../../../shared/dumb_components/member_fields.js'
 import { reduxForm } from 'redux-form'
-import { validate, fieldStructure, required, removed } from '../../../admin/form_fields/member.js'
+import { validate, fieldStructure, required, removed, read_only_user }
+  from '../../../shared/form_fields/member.js'
 import buttons from '../my_details/buttons.js'
 
 import { reject, contains, concat, __ } from 'ramda'
 
 export default ({ toggle_member_mode, mode, update_member_user, active_tab, ...props }) =>
-{ console.log(filter_fields(removed));
-  return(  <ViewMember
+  <ViewMember
     {...props}
     fields={filter_fields(removed)(field_mapper[active_tab])}
     Buttons={buttons}
@@ -21,7 +21,8 @@ export default ({ toggle_member_mode, mode, update_member_user, active_tab, ...p
     mode={mode}
     onSubmit={update_member_user}
     required={required}
-  />)}
+    read_only={read_only_user}
+  />
 
 const ViewMember = reduxForm(
   { form: 'user'
@@ -33,7 +34,7 @@ const ViewMember = reduxForm(
 const filter_fields = to_be_removed => reject(contains(__, to_be_removed))
 
 const field_mapper =
-  { 'contact_details': concat(fieldStructure.address, fieldStructure.personal)
-  , 'membership_details': concat(['id'], fieldStructure.membership)
-  , 'account_details': concat(['id'], fieldStructure.edit)
+  { contact_details: concat(fieldStructure.address, fieldStructure.personal)
+  , membership_details: concat(['id'], fieldStructure.membership)
+  , account_details: concat(['id'], fieldStructure.edit)
   }
