+++ src/user/components/my_details/edit_details.js
@@ -1,7 +1,8 @@
 import React from 'react'
-import MemberFields from '../../../admin/dumb_components/member_fields.js'
+import MemberFields from '../../../shared/dumb_components/member_fields.js'
 import { reduxForm } from 'redux-form'
-import { validate, fieldStructure, required, removed } from '../../../admin/form_fields/member.js'
+import { validate, fieldStructure, required, removed, read_only_user }
+  from '../../../shared/form_fields/member.js'
 import buttons from '../my_details/buttons.js'
 
 import { reject, contains, concat, __ } from 'ramda'
@@ -20,6 +21,7 @@ export default ({ toggle_member_mode, mode, update_member_user, active_tab, ...p
     mode={mode}
     onSubmit={update_member_user}
     required={required}
+    read_only={read_only_user}
   />
 
 const ViewMember = reduxForm(
@@ -32,7 +34,7 @@ const ViewMember = reduxForm(
 const filter_fields = to_be_removed => reject(contains(__, to_be_removed))
 
 const field_mapper =
-  { 'contact_details': concat(fieldStructure.address, fieldStructure.personal)
-  , 'membership_details': concat(['id'], fieldStructure.membership)
-  , 'account_details': concat(['id'], fieldStructure.edit)
+  { contact_details: concat(fieldStructure.address, fieldStructure.personal)
+  , membership_details: concat(['id'], fieldStructure.membership)
+  , account_details: concat(['id'], fieldStructure.edit)
   }
