+++ src/user/redux/form_reducer.js
@@ -1,7 +1,7 @@
 /* @flow */
 const { reducer: form } = require('redux-form')
 import member_reducer from '../../shared/redux/modules/member.js'
-import { normalise as member } from '../../admin/form_fields/member.js'
+import { normalise as member } from '../../shared/form_fields/member.js'
 
 export default form.plugin(
   { user: member_reducer
